+++ packages/artrelay_flutter/lib/artrelay_flutter.dart
@@ -6,6 +6,7 @@ export 'package:artrelay_core/artrelay_core.dart';
 
 export 'src/app_links_callback_source.dart';
 export 'src/artrelay_oauth_client.dart';
+export 'src/background_transfer_manager.dart' show BackgroundTransferManager;
 export 'src/secure_pending_authorization_store.dart';
 export 'src/secure_token_store.dart';
 export 'src/system_uri_launcher.dart';